requirements.txt
django
psycopg2-binary
django-environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular